// web/src/api/generated/models/ApiV1ExecutionStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api/v1/user.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ApiV1ExecutionStatus = {
    ExecutionStatusUnspecified: 'EXECUTION_STATUS_UNSPECIFIED',
    ExecutionStatusPending: 'EXECUTION_STATUS_PENDING',
    ExecutionStatusRunning: 'EXECUTION_STATUS_RUNNING',
    ExecutionStatusCompleted: 'EXECUTION_STATUS_COMPLETED',
    ExecutionStatusFailed: 'EXECUTION_STATUS_FAILED',
    ExecutionStatusCancelled: 'EXECUTION_STATUS_CANCELLED'
} as const;
export type ApiV1ExecutionStatus = typeof ApiV1ExecutionStatus[keyof typeof ApiV1ExecutionStatus];


export function instanceOfApiV1ExecutionStatus(value: any): boolean {
    for (const key in ApiV1ExecutionStatus) {
        if (Object.prototype.hasOwnProperty.call(ApiV1ExecutionStatus, key)) {
            if (ApiV1ExecutionStatus[key as keyof typeof ApiV1ExecutionStatus] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ApiV1ExecutionStatusFromJSON(json: any): ApiV1ExecutionStatus {
    return ApiV1ExecutionStatusFromJSONTyped(json, false);
}

export function ApiV1ExecutionStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApiV1ExecutionStatus {
    return json as ApiV1ExecutionStatus;
}

export function ApiV1ExecutionStatusToJSON(value?: ApiV1ExecutionStatus | null): any {
    return value as any;
}

export function ApiV1ExecutionStatusToJSONTyped(value: any, ignoreDiscriminator: boolean): ApiV1ExecutionStatus {
    return value as ApiV1ExecutionStatus;
}